chore: improve List exports
